+++ packages/react/template/base/src/utils/test-utils.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import type { PreloadedState } from '@reduxjs/toolkit';
+import ReactTestRenderer from 'react-test-renderer';
+import { Provider } from 'react-redux';
+
+import { createStore } from '@/store/';
+import type { AppStore, RootState } from '@/store/';
+
+// This type interface extends the default options for render from RTL, as well
+// as allows the user to specify other things such as initialState, store.
+interface ExtendedRenderOptions {
+  preloadedState?: PreloadedState<RootState>
+  store?: AppStore
+}
+
+export function renderWithProviders(
+  component: React.ReactElement,
+  {
+    preloadedState,
+    // Automatically create a store instance if no store was passed in
+    store = createStore(preloadedState),
+  }: ExtendedRenderOptions = {},
+) {
+  function Wrapper(): JSX.Element {
+    return <Provider store={store}>{ component }</Provider>;
+  }
+
+  return ReactTestRenderer.create(<Wrapper />);
+}
+
+export function findByClassName(root: ReactTestRenderer.ReactTestRendererJSON, className: string) {
+  let nextNodes: ReactTestRenderer.ReactTestRendererNode[] | undefined;
+  const waitingChildren: ReactTestRenderer.ReactTestRendererNode[][] = [[root]];
+  const regClassName = new RegExp(`\\b${className}\\b`);
+  do {
+    nextNodes = waitingChildren.shift();
+    if (!nextNodes) continue;
+    for (const node of nextNodes) {
+      if (typeof node === 'string') continue;
+      if (node.props?.className?.match(regClassName)) return node;
+      if (node.children) {
+        waitingChildren.push(node.children);
+      }
+    }
+  } while (waitingChildren.length > 0);
+  return null;
+}
+