+++ packages/shared/src/merge-config.ts
@@ -1,65 +1,82 @@
-function isObject(value: unknown): boolean {
-  return Object.prototype.toString.call(value) === '[object Object]';
-}
-// 为了解决 JSONObject 等无穷循环的类型导致报错，这里只支持最多 10 层 deep partial
-// Type instantiation is excessively deep and possibly infinite
 type DeepPartial<T> = T extends object ? {
-  [P in keyof T]?: DeepPartial10<T[P]>;
+  [P in keyof T]?: T[P] extends object ? DeepPartial<T[P]>: T[P];
 } : T;
 
-type DeepPartial10<T> = T extends object ? { [P in keyof T]?: DeepPartial9<T[P]>; } : T;
-type DeepPartial9<T> = T extends object ? { [P in keyof T]?: DeepPartial8<T[P]>; } : T;
-type DeepPartial8<T> = T extends object ? { [P in keyof T]?: DeepPartial7<T[P]>; } : T;
-type DeepPartial7<T> = T extends object ? { [P in keyof T]?: DeepPartial6<T[P]>; } : T;
-type DeepPartial6<T> = T extends object ? { [P in keyof T]?: DeepPartial5<T[P]>; } : T;
-type DeepPartial5<T> = T extends object ? { [P in keyof T]?: DeepPartial4<T[P]>; } : T;
-type DeepPartial4<T> = T extends object ? { [P in keyof T]?: DeepPartial3<T[P]>; } : T;
-type DeepPartial3<T> = T extends object ? { [P in keyof T]?: DeepPartial2<T[P]>; } : T;
-type DeepPartial2<T> = T extends object ? Partial<T> : T;
-export default function mergeConfig<T>(
+type MergeResult = {
+  merged: true,
+  // 合并结果可以包含任何类型的值，这里要更严谨可以用泛型
+  value: any,
+} | { merged: false };
+
+const isObject = (value: unknown): value is object => value !== null
+  && Object.prototype.toString.call(value) === '[object Object]';
+
+const mergeWith = (value: any): MergeResult => ({
+  merged: true,
+  value,
+});
+
+const pass = (): MergeResult => ({ merged: false });
+
+const simpleMerge = (existing: unknown, newValue: unknown): MergeResult => {
+  if (newValue === undefined) return mergeWith(existing);
+  if (existing === null || existing === undefined) return mergeWith(newValue);
+  return pass();
+};
+
+// array merge 只关注是否参数是 2 个 array，具体类型不关心
+const arrayMerge = (existing: unknown, newValue: unknown): MergeResult => (
+  (!Array.isArray(newValue) || !Array.isArray(existing))
+    ? pass()
+    : mergeWith([...existing, ...newValue])
+);
+
+// 如果 2个值都是 object，则递归调用 mergeConfig
+const objectMerge = (existing: unknown, newValue: unknown): MergeResult => (
+  (!isObject(existing) || !isObject(newValue))
+    ? pass()
+    : mergeWith(mergeConfig(existing, newValue))
+);
+
+// 如果 2 个值类型相同，直接覆盖
+const sameTypeMerge = (
+  existing: unknown,
+  newValue: unknown,
+): MergeResult => (
+  typeof existing === typeof newValue ? mergeWith(newValue) : pass()
+);
+
+const useMergedValueOrExisting = (
+  mergeResult: MergeResult,
+  existing: unknown,
+) => (mergeResult.merged ? mergeResult.value : existing);
+
+const MERGE_METHODS = [simpleMerge, arrayMerge, objectMerge, sameTypeMerge];
+
+const tryAllMergeMethods = (
+  existing: unknown,
+  newValue: unknown,
+): MergeResult => MERGE_METHODS.reduce<MergeResult>((mergeResult, fn) => {
+  if (mergeResult.merged) return mergeResult;
+  return fn(existing, newValue);
+}, { merged: false });
+
+const mergeOneConfig = <T extends object>(
+  defaults: T,
+  override: DeepPartial<T>,
+): T => (Object.keys(override) as Array<keyof DeepPartial<T>>)
+    .reduce((target, key) => ({
+      ...target,
+      [key as keyof T]:
+        useMergedValueOrExisting(
+          tryAllMergeMethods(target[key as keyof T], override[key]),
+          target[key as keyof T],
+        ),
+    }), defaults);
+
+const mergeConfig = <T extends object>(
   defaults: T,
   ...overrides: Array<DeepPartial<T>>
-): T {
-  const merged = { ...defaults };
+): T => overrides.reduce(mergeOneConfig, defaults);
 
-  overrides.forEach((override) => {
-    (Object.keys(override as {}) as Array<keyof DeepPartial<T>>).forEach((key) => {
-      const value = override[key] as unknown as T[keyof T];
-
-      if (value === undefined) {
-        return;
-      }
-
-      const existing = merged[key as keyof T];
-
-      if (existing === null || existing === undefined) {
-        merged[key as keyof T] = value;
-        return;
-      }
-
-      if (Array.isArray(value)) {
-        if (Array.isArray(existing)) {
-          // isArray 添加了 any 属性导致类型转换需要先转 unknown
-          const mergedArray = [...existing, ...value] as unknown as T[keyof T];
-          merged[key as keyof T] = mergedArray;
-        }
-        return;
-      }
-
-      if (isObject(existing)) {
-        if (isObject(value) && value !== null) {
-          merged[key as keyof T] = mergeConfig(
-            existing,
-            value as DeepPartial<T[keyof T]>,
-          );
-        }
-        return;
-      }
-
-      if (typeof value === typeof existing) {
-        merged[key as keyof T] = value;
-      }
-    });
-  });
-  return merged;
-}
+export default mergeConfig;